+++ HealthCheck/HealthCheck/src/app/app.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { ConnectionService } from 'angular-connection-service';
 
 @Component({
   selector: 'app-root',
@@ -7,4 +8,19 @@ import { Component } from '@angular/core';
 })
 export class AppComponent {
   title = 'HealthCheck';
+
+  hasNetworkConnection: boolean = true;
+  hasInternetAccess: boolean = true;
+
+  constructor(private connectionService: ConnectionService) {
+    this.connectionService.monitor().subscribe((currentState: any) => {
+      this.hasNetworkConnection = currentState.hasNetworkConnection;
+      this.hasInternetAccess = currentState.HasInternetAccess;
+    });
+  }
+
+  public isOnline() {
+    return this.hasNetworkConnection && this.hasInternetAccess;
+  }
+
 }