+++ HealthCheck/HealthCheckAPI/Program.cs
@@ -1,4 +1,5 @@
 using HealthCheckAPI.HealthCheck;
+using Microsoft.AspNetCore.Cors;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -13,6 +14,12 @@ builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
+
+builder.Services.AddCors(options => options.AddPolicy(name: "AngularPolicy", cfg => {
+  cfg.AllowAnyHeader();
+  cfg.AllowAnyMethod();
+  cfg.WithOrigins(builder.Configuration["AllowedCORS"]);
+}));
 
 var app = builder.Build();
 
@@ -33,6 +40,8 @@ app.UseAuthorization();
 app.UseHealthChecks(new PathString("/api/health"), new CustomHealthCheckOptions());
 
 app.MapControllers();
+
+app.MapMethods("/api/heartbeat", new[] { "HEAD" }, () => Results.Ok());
 
 app.MapFallbackToFile("/index.html");